% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actg019-data.R
\docType{data}
\name{actg019}
\alias{actg019}
\title{ACTG 019 Data}
\format{
A data frame with 862 rows and 4 variables:
\describe{
\item{age}{age, in years}
\item{treatment}{treatment indicator, 0 = placebo, 1 = AZT}
\item{race}{race indicator, 0 = non-white, 1 = white}
\item{cd4}{CD4 count}
}
}
\description{
ACTG 019 Data
}
\references{
\url{https://clinicaltrials.gov/ct2/show/NCT00000736}
}
\author{
Joseph Ibrahim \email{ibrahim@bios.unc.edu}
}
\keyword{data}
