% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actg036-data.R
\docType{data}
\name{actg036}
\alias{actg036}
\title{ACTG 036 Data}
\format{
A data frame with 183 rows and 4 variables:
\describe{
\item{outcome}{outcome variable with 1 indicating death, development of AIDS or ARC and 0 otherwise}
\item{age}{age, in years}
\item{treatment}{treatment indicator, 0 = placebo, 1 = AZT}
\item{race}{race indicator, 0 = non-white, 1 = white}
\item{cd4}{CD4 count}
}
}
\description{
ACTG 036 Data
}
\references{
\url{https://clinicaltrials.gov/ct2/show/NCT00001104}
}
\author{
Joseph Ibrahim \email{ibrahim@bios.unc.edu}
}
\keyword{data}
