% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aft_commensurate.R
\name{aft.commensurate}
\alias{aft.commensurate}
\title{Posterior of commensurate prior (CP)}
\usage{
aft.commensurate(
  formula,
  data.list,
  dist = "weibull",
  beta0.mean = NULL,
  beta0.sd = NULL,
  p.spike = 0.1,
  spike.mean = 200,
  spike.sd = 0.1,
  slab.mean = 0,
  slab.sd = 5,
  scale.mean = NULL,
  scale.sd = NULL,
  get.loglik = FALSE,
  iter_warmup = 1000,
  iter_sampling = 1000,
  chains = 4,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates.
The response is a survival object as returned by the \code{survival::Surv(time, event)} function,
where event is a binary indicator for event (0 = no event, 1 = event has occurred). The type of
censoring is assumed to be right-censoring.}

\item{data.list}{a list of \code{data.frame}s. The first element in the list is the current data, and the rest
are the historical data sets. For fitting accelerated failure time (AFT) models, all historical
data sets will be stacked into one historical data set.}

\item{dist}{a character indicating the distribution of survival times. Currently, \code{dist} can be one of the
following values: "weibull", "lognormal", or "loglogistic". Defaults to "weibull".}

\item{beta0.mean}{a scalar or a vector whose dimension is equal to the number of regression coefficients
giving the mean parameters for the prior on the historical data regression coefficients. If a
scalar is provided, \code{beta0.mean} will be a vector of repeated elements of the given scalar.
Defaults to a vector of 0s.}

\item{beta0.sd}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the sd parameters for the prior on the historical data regression coefficients. If a scalar is
provided, same as for \code{beta0.mean}. Defaults to a vector of 10s.}

\item{p.spike}{a scalar between 0 and 1 giving the probability of the spike component in spike-and-slab prior
on commensurability parameter \eqn{\tau}. Defaults to 0.1.}

\item{spike.mean}{a scalar giving the location parameter for the half-normal prior (spike component) on \eqn{\tau}.
Defaults to 200.}

\item{spike.sd}{a scalar giving the scale parameter for the half-normal prior (spike component) on \eqn{\tau}.
Defaults to 0.1.}

\item{slab.mean}{a scalar giving the location parameter for the half-normal prior (slab component) on \eqn{\tau}.
Defaults to 0.}

\item{slab.sd}{a scalar giving the scale parameter for the half-normal prior (slab component) on \eqn{\tau}.
Defaults to 5.}

\item{scale.mean}{location parameter for the half-normal prior on the scale parameters of current and historical
data models. Defaults to 0.}

\item{scale.sd}{scale parameter for the half-normal prior on the scale parameters of current and historical data
models. Defaults to 10.}

\item{get.loglik}{whether to generate log-likelihood matrix. Defaults to FALSE.}

\item{iter_warmup}{number of warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_warmup} in
\code{sample()} method in cmdstanr package.}

\item{iter_sampling}{number of post-warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_sampling}
in \code{sample()} method in cmdstanr package.}

\item{chains}{number of Markov chains to run. Defaults to 4. See the argument \code{chains} in \code{sample()} method in
cmdstanr package.}

\item{...}{arguments passed to \code{sample()} method in cmdstanr package (e.g., \code{seed}, \code{refresh}, \code{init}).}
}
\value{
The function returns an object of class \code{draws_df} giving posterior samples, with an attribute called 'data' which includes
the list of variables specified in the data block of the Stan program.
}
\description{
Sample from the posterior distribution of an accelerated failure time (AFT) model using the commensurate prior (CP)
by Hobbs et al. (2011) \url{doi:10.1111/j.1541-0420.2011.01564.x}.
}
\details{
The commensurate prior (CP) assumes that the regression coefficients for the current data model conditional on those
for the historical data model are independent normal distributions with mean equal to the corresponding regression
coefficients for the historical data and variance equal to the inverse of the corresponding elements of a vector of
precision parameters (referred to as the commensurability parameter \eqn{\tau}). We regard \eqn{\tau} as random and elicit
a spike-and-slab prior, which is specified as a mixture of two half-normal priors, on \eqn{\tau}.

The number of current data regression coefficients is assumed to be the same as that of historical data regression
coefficients. The scale parameters for both current and historical data models are assumed to be independent and
identically distributed, each assigned a half-normal prior.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  if(requireNamespace("survival")){
    library(survival)
    data(E1684)
    data(E1690)
    ## take subset for speed purposes
    E1684 = E1684[1:100, ]
    E1690 = E1690[1:50, ]
    ## replace 0 failure times with 0.50 days
    E1684$failtime[E1684$failtime == 0] = 0.50/365.25
    E1690$failtime[E1690$failtime == 0] = 0.50/365.25
    E1684$cage = as.numeric(scale(E1684$age))
    E1690$cage = as.numeric(scale(E1690$age))
    data_list = list(currdata = E1690, histdata = E1684)
    aft.commensurate(
      formula = survival::Surv(failtime, failcens) ~ treatment + sex + cage + node_bin,
      data.list = data_list,
      dist = "weibull",
      p.spike = 0.1,
      chains = 1, iter_warmup = 500, iter_sampling = 1000
    )
  }
}
}
\references{
Hobbs, B. P., Carlin, B. P., Mandrekar, S. J., and Sargent, D. J. (2011). Hierarchical commensurate and power prior models for adaptive incorporation of historical information in clinical trials. Biometrics, 67(3), 1047–1056.
}
