% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{data.checks}
\alias{data.checks}
\title{check if the input data are in appropriate forms}
\usage{
data.checks(formula, family, data.list, offset.list)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates.}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}.}

\item{data.list}{a list of \code{data.frame}s. The first element in the list is the current data, and the rest
are the historical datasets.}

\item{offset.list}{a list of vectors giving the offsets for each data. The length of offset.list is equal to
the length of data.list. The length of each element of offset.list is equal to the number
of rows in the corresponding element of data.list. Defaults to a list of vectors of 0s.}
}
\description{
check if the input data are in appropriate forms
}
