% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfam_loglik.R
\name{gamma_glm_lp}
\alias{gamma_glm_lp}
\title{compute density for gamma GLM}
\usage{
gamma_glm_lp(y, beta, X, link, offs, phi)
}
\arguments{
\item{y}{response vector}

\item{beta}{regression coefficients}

\item{X}{design matrix}

\item{link}{integer giving link function}

\item{offs}{offset}

\item{phi}{dispersion parameter}
}
\description{
compute density for gamma GLM
}
