% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{get.dist.link}
\alias{get.dist.link}
\title{get distribution and link for a GLM}
\usage{
get.dist.link(family)
}
\arguments{
\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}.}
}
\description{
get distribution and link for a GLM
}
