% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfam_loglik.R
\name{get_lp2mean}
\alias{get_lp2mean}
\title{Compute mean from linear predictor in a GLM}
\usage{
get_lp2mean(eta, link)
}
\arguments{
\item{eta}{linear predictor}

\item{link}{integer giving link function}
}
\description{
Compute mean from linear predictor in a GLM
}
