% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_bhm_lognc.R
\name{glm.bhm.lognc}
\alias{glm.bhm.lognc}
\title{Estimate the logarithm of the normalizing constant for Bayesian hierarchical model (BHM)}
\usage{
glm.bhm.lognc(post.samples, bridge.args = NULL)
}
\arguments{
\item{post.samples}{output from \code{\link[=glm.bhm]{glm.bhm()}} giving posterior samples of a GLM under the Bayesian hierarchical
model (BHM) or samples from the prior induced by the BHM, with an attribute called 'data'
which includes the list of variables specified in the data block of the Stan program.}

\item{bridge.args}{a \code{list} giving arguments (other than \code{samples}, \code{log_posterior}, \code{data}, \code{lb}, and \code{ub})
to pass onto \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}.}
}
\value{
The function returns a \code{list} with the following objects

\describe{
\item{lognc}{the estimated logarithm of the normalizing constant}

\item{bs}{an object of class \code{bridge} or \code{bridge_list} giving the output from \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}}
}
}
\description{
Uses bridge sampling to estimate the logarithm of the normalizing constant for Bayesian hierarchical
model (BHM) using all data sets or using historical data sets only.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  data(actg019)
  data(actg036)
  ## take subset for speed purposes
  actg019 = actg019[1:100, ]
  actg036 = actg036[1:50, ]
  formula = outcome ~ scale(age) + race + treatment + scale(cd4)
  family = binomial('logit')
  data_list = list(currdata = actg019, histdata = actg036)
  d.bhm = glm.bhm(
    formula = formula,
    family = family,
    data.list = data_list,
    chains = 1, iter_warmup = 500, iter_sampling = 1000
  )
  glm.bhm.lognc(
    post.samples = d.bhm,
    bridge.args = list(silent = TRUE)
  )
}
}
\references{
Gronau, Q. F., Singmann, H., and Wagenmakers, E.-J. (2020). bridgesampling: An r package for estimating normalizing constants. Journal of Statistical Software, 92(10).
}
