% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_bhm.R
\name{glm.bhm}
\alias{glm.bhm}
\title{Bayesian hierarchical model}
\usage{
glm.bhm(
  formula,
  family,
  data.list,
  offset.list = NULL,
  meta.mean.mean = NULL,
  meta.mean.sd = NULL,
  meta.sd.mean = NULL,
  meta.sd.sd = NULL,
  disp.mean = NULL,
  disp.sd = NULL,
  local.location = NULL,
  iter_warmup = 1000,
  iter_sampling = 1000,
  chains = 4,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates.}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}.}

\item{data.list}{a list of \code{data.frame}s. The first element in the list is the current data, and the rest
are the historical datasets.}

\item{offset.list}{a list of vectors giving the offsets for each data. The length of offset.list is equal to
the length of data.list. The length of each element of offset.list is equal to the number
of rows in the corresponding element of data.list. Defaults to a list of vectors of 0s.}

\item{meta.mean.mean}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the means for the normal hyperpriors on the mean hyperparameters of regression coefficients. If
a scalar is provided, meta.mean.mean will be a vector of repeated elements of the given scalar.
Defaults to a vector of 0s.}

\item{meta.mean.sd}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the sds for the normal hyperpriors on the mean hyperparameters of regression coefficients. If a
scalar is provided, same as for meta.mean.mean. Defaults to a vector of 10s.}

\item{meta.sd.mean}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the means for the half-normal hyperpriors on the sd hyperparameters of regression coefficients.
If a scalar is provided, same as for meta.mean.mean. Defaults to a vector of 0s.}

\item{meta.sd.sd}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the sds for the half-normal hyperpriors on the sd hyperparameters of regression coefficients.
If a scalar is provided, same as for meta.mean.mean. Defaults to a vector of 1s.}

\item{disp.mean}{a scalar or a vector whose dimension is equal to the number of datasets (including the current
data) giving the means for the half-normal hyperpriors on the dispersion parameters. If a scalar
is provided, same as for meta.mean.mean. Defaults to a vector of 0s.}

\item{disp.sd}{a scalar or a vector whose dimension is equal to the number of datasets (including the current
data) giving the sds for the half-normal hyperpriors on the dispersion parameters. If a scalar
is provided, same as for meta.mean.mean. Defaults to a vector of 10s.}

\item{local.location}{a file path giving the desired location of the local copies of all the .stan model files in the
package. Defaults to the path created by \code{rappdirs::user_cache_dir("hdbayes")}.}

\item{iter_warmup}{number of warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_warmup} in
\code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{iter_sampling}{number of post-warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_sampling}
in \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{chains}{number of Markov chains to run. Defaults to 4. See the argument \code{chains} in \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{...}{arguments passed to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}} (e.g. seed, refresh, init).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{             an object of class `draws_df` giving posterior samples
}\if{html}{\out{</div>}}
}
\description{
Sample from the posterior distribution of a GLM using the Bayesian Hierarchical Model (BHM). This model assumes that
the regression coefficients of the historical and current data are different, but are correlated through a common
distribution, whose hyperparameters (i.e., mean and standard deviation (the covariance is assumed to have a diagonal
structure)) are treated as random.
The mean hyperparmaters are independent normal and the standard deviation (sd) hyperparameters are independent half-normal.
The dispersion parameters (if applicable) are independent.
}
\examples{
data(actg019)
data(actg036)
## take subset for speed purposes
actg019 = actg019[1:200, ]
actg036 = actg036[1:100, ]
data_list = list(currdata = actg019, histdata = actg036)
glm.bhm(
  formula = cd4 ~ treatment + age + race,
  family = poisson('log'),
  data.list = data_list,
  chains = 1, iter_warmup = 500, iter_sampling = 1000
)
}
