% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_commensurate_lognc.R
\name{glm.commensurate.lognc}
\alias{glm.commensurate.lognc}
\title{Estimate the logarithm of the normalizing constant for commensurate prior (CP)}
\usage{
glm.commensurate.lognc(post.samples, bridge.args = NULL)
}
\arguments{
\item{post.samples}{output from \code{\link[=glm.commensurate]{glm.commensurate()}} giving posterior samples of a GLM under the commensurate
prior (CP), with an attribute called 'data' which includes the list of variables specified
in the data block of the Stan program.}

\item{bridge.args}{a \code{list} giving arguments (other than \code{samples}, \code{log_posterior}, \code{data}, \code{lb}, and \code{ub})
to pass onto \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}.}
}
\value{
The function returns a \code{list} with the following objects

\describe{
\item{lognc}{the estimated logarithm of the normalizing constant}

\item{bs}{an object of class \code{bridge} or \code{bridge_list} giving the output from \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}}
}
}
\description{
Uses bridge sampling to estimate the logarithm of the normalizing constant for the commensurate prior (CP) using
historical data sets.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  data(actg036)
  ## take subset for speed purposes
  actg036 = actg036[1:50, ]
  hist.data.list = list(histdata = actg036)
  standat = hdbayes:::get.stan.data.cp(
    formula = cd4 ~ treatment + age + race,
    family = poisson(),
    data.list = hist.data.list
  )
  glm_commensurate_prior = instantiate::stan_package_model(
    name = "glm_commensurate_prior",
    package = "hdbayes"
  )
  fit = glm_commensurate_prior$sample(
    data = standat,
    iter_warmup = 500, iter_sampling = 1000, chains = 1
  )
  d.cp.hist  = fit$draws(format = 'draws_df')
  attr(x = d.cp.hist, which = 'data') = standat
  glm.commensurate.lognc(
    post.samples = d.cp.hist
  )
}
}
\references{
Hobbs, B. P., Carlin, B. P., Mandrekar, S. J., and Sargent, D. J. (2011). Hierarchical commensurate and power prior models for adaptive incorporation of historical information in clinical trials. Biometrics, 67(3), 1047–1056.

Gronau, Q. F., Singmann, H., and Wagenmakers, E.-J. (2020). bridgesampling: An r package for estimating normalizing constants. Journal of Statistical Software, 92(10).
}
