% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_cp_lognc.R
\name{glm.cp.lognc}
\alias{glm.cp.lognc}
\title{Estimate the logarithm of the normalizing constant for commensurate prior (CP)}
\usage{
glm.cp.lognc(
  formula,
  family,
  hist.data.list,
  tau,
  hist.offset.list = NULL,
  beta0.mean = NULL,
  beta0.sd = NULL,
  hist.disp.mean = NULL,
  hist.disp.sd = NULL,
  bridge.args = NULL,
  iter_warmup = 1000,
  iter_sampling = 1000,
  chains = 4,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}}

\item{hist.data.list}{a list of \code{data.frame}s giving historical data sets.}

\item{tau}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the commensurate prior parameters. If a scalar is provided, tau will be a vector of repeated
elements of the given scalar. Each element of tau must be positive, corresponding to a normal
precision parameter.}

\item{hist.offset.list}{a list of vectors giving the offsets for each historical data. The length of hist.offset.list
is equal to the length of hist.data.list. The length of each element of offset.list is equal to
the number of rows in the corresponding element of hist.data.list. Defaults to a list of vectors
of 0s.}

\item{beta0.mean}{a scalar or a vector whose dimension is equal to the number of regression coefficients
giving the mean parameters for the prior on the historical data regression coefficients. If a
scalar is provided, same as for tau. Defaults to a vector of 0s.}

\item{beta0.sd}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the sd parameters for the prior on the historical data regression coefficients. If a scalar is
provided, same as for tau. Defaults to a vector of 10s.}

\item{hist.disp.mean}{a scalar or a vector whose dimension is equal to the number of historical data sets giving the
means for the half-normal priors on the dispersion parameters. If a scalar is provided, same as
for tau. Defaults to a vector of 0s.}

\item{hist.disp.sd}{a scalar or a vector whose dimension is equal to the number of historical data sets giving the
sds for the half-normal priors on the dispersion parameters. If a scalar is provided, same as
for tau. Defaults to a vector of 10s.}

\item{bridge.args}{a \code{list} giving arguments (other than samples, log_posterior, data, lb, ub) to pass
onto \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}.}

\item{iter_warmup}{number of warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_warmup} in
\code{sample()} method in cmdstanr package.}

\item{iter_sampling}{number of post-warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_sampling}
in \code{sample()} method in cmdstanr package.}

\item{chains}{number of Markov chains to run. Defaults to 4. See the argument \code{chains} in \code{sample()} method in
cmdstanr package.}

\item{...}{arguments passed to \code{sample()} method in cmdstanr package (e.g. seed, refresh, init).}
}
\value{
The function returns a \code{list} with the following objects

\describe{
\item{lognc}{the estimated logarithm of the normalizing constant}

\item{min_ess_bulk}{the minimum estimated bulk effective sample size of the MCMC sampling}

\item{max_Rhat}{the maximum Rhat}

\item{bs}{an object of class \code{bridge} or \code{bridge_list} giving the output from \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}}
}
}
\description{
Uses Markov chain Monte Carlo (MCMC) and bridge sampling to estimate the logarithm of the normalizing
constant for the CP using historical data sets.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  data(actg036)
  ## take subset for speed purposes
  actg036 = actg036[1:50, ]
  glm.cp.lognc(
    formula = cd4 ~ treatment + age + race,
    family = poisson(),
    hist.data.list = list(histdata = actg036),
    tau = 5,
    chains = 1, iter_warmup = 500, iter_sampling = 2000
  )
}
}
\references{
Hobbs, B. P., Carlin, B. P., Mandrekar, S. J., and Sargent, D. J. (2011). Hierarchical commensurate and power prior models for adaptive incorporation of historical information in clinical trials. Biometrics, 67(3), 1047–1056.

Gronau, Q. F., Singmann, H., and Wagenmakers, E.-J. (2020). bridgesampling: An r package for estimating normalizing constants. Journal of Statistical Software, 92(10).
}
