% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_logml.R
\name{glm.logml.bhm}
\alias{glm.logml.bhm}
\title{Log marginal likelihood of a GLM under Bayesian hierarchical model (BHM)}
\usage{
glm.logml.bhm(
  post.samples,
  formula,
  family,
  data.list,
  offset.list = NULL,
  meta.mean.mean = NULL,
  meta.mean.sd = NULL,
  meta.sd.mean = NULL,
  meta.sd.sd = NULL,
  disp.mean = NULL,
  disp.sd = NULL,
  bridge.args = NULL,
  iter_warmup = 1000,
  iter_sampling = 1000,
  chains = 4,
  ...
)
}
\arguments{
\item{post.samples}{an object of class \code{draws_df}, \code{draws_matrix}, \code{matrix}, or \code{data.frame} giving posterior
samples of a GLM using the PP, such as the output from \code{\link[=glm.pp]{glm.pp()}}. Each row corresponds to the
posterior samples obtained from one iteration of MCMC. The column names of \code{post.samples} should
include the names of covariates for regression coefficients, such as "(Intercept)", and
"dispersion" for the dispersion parameter, if applicable.}

\item{formula}{a two-sided formula giving the relationship between the response variable and covariates.}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}.}

\item{data.list}{a list of \code{data.frame}s. The first element in the list is the current data, and the rest
are the historical data sets.}

\item{offset.list}{a list of vectors giving the offsets for each data. The length of offset.list is equal to
the length of data.list. The length of each element of offset.list is equal to the number
of rows in the corresponding element of data.list. Defaults to a list of vectors of 0s.}

\item{meta.mean.mean}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the means for the normal hyperpriors on the mean hyperparameters of regression coefficients. If
a scalar is provided, meta.mean.mean will be a vector of repeated elements of the given scalar.
Defaults to a vector of 0s.}

\item{meta.mean.sd}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the sds for the normal hyperpriors on the mean hyperparameters of regression coefficients. If a
scalar is provided, same as for meta.mean.mean. Defaults to a vector of 10s.}

\item{meta.sd.mean}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the means for the half-normal hyperpriors on the sd hyperparameters of regression coefficients.
If a scalar is provided, same as for meta.mean.mean. Defaults to a vector of 0s.}

\item{meta.sd.sd}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the sds for the half-normal hyperpriors on the sd hyperparameters of regression coefficients.
If a scalar is provided, same as for meta.mean.mean. Defaults to a vector of 1s.}

\item{disp.mean}{a scalar or a vector whose dimension is equal to the number of data sets (including the current
data) giving the means for the half-normal priors on the dispersion parameters. If a scalar is
provided, same as for meta.mean.mean. Defaults to a vector of 0s.}

\item{disp.sd}{a scalar or a vector whose dimension is equal to the number of data sets (including the current
data) giving the sds for the half-normal priors on the dispersion parameters. If a scalar is
provided, same as for meta.mean.mean. Defaults to a vector of 10s.}

\item{bridge.args}{a \code{list} giving arguments (other than samples, log_posterior, data, lb, ub) to pass
onto \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}.}

\item{iter_warmup}{number of warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_warmup} in
\code{sample()} method in cmdstanr package.}

\item{iter_sampling}{number of post-warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_sampling}
in \code{sample()} method in cmdstanr package.}

\item{chains}{number of Markov chains to run. Defaults to 4. See the argument \code{chains} in \code{sample()} method in
cmdstanr package.}

\item{...}{arguments passed to \code{sample()} method in cmdstanr package (e.g. seed, refresh, init).}
}
\value{
The function returns a \code{list} with the following objects

\describe{
\item{model}{"BHM"}

\item{logml}{the estimated logarithm of the marginal likelihood}
}
}
\description{
Uses Markov chain Monte Carlo (MCMC) and bridge sampling to estimate the logarithm of the marginal
likelihood of a GLM under BHM. The function's arguments mirror those of the function \code{\link[=glm.bhm]{glm.bhm()}},
including two additional arguments: \code{post.samples} and \code{bridge.args}. These provide posterior samples
from the GLM under BHM (for example, the output from \code{\link[=glm.bhm]{glm.bhm()}}) and specify arguments to pass onto
\code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}} (other than \code{samples}, \code{log_posterior}, \code{data}, \code{lb}, \code{ub}), respectively.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  data(actg019)
  data(actg036)
  ## take subset for speed purposes
  actg019 = actg019[1:100, ]
  actg036 = actg036[1:50, ]
  formula = outcome ~ scale(age) + race + treatment + scale(cd4)
  family = binomial('logit')
  data_list = list(currdata = actg019, histdata = actg036)
  d.bhm = glm.bhm(
    formula = formula,
    family = family,
    data.list = data_list,
    chains = 1, iter_warmup = 500, iter_sampling = 1000
  )
  glm.logml.bhm(
    post.samples = d.bhm,
    formula = formula, family = family,
    data.list = data_list,
    chains = 1, iter_warmup = 500, iter_sampling = 1000
  )
}
}
\references{
Gronau, Q. F., Singmann, H., and Wagenmakers, E.-J. (2020). bridgesampling: An r package for estimating normalizing constants. Journal of Statistical Software, 92(10).
}
