% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_logml_napp.R
\name{glm.logml.napp}
\alias{glm.logml.napp}
\title{Log marginal likelihood of a GLM under normalized asymptotic power prior (NAPP)}
\usage{
glm.logml.napp(
  post.samples,
  formula,
  family,
  data.list,
  offset.list = NULL,
  a0.shape1 = 1,
  a0.shape2 = 1,
  bridge.args = NULL
)
}
\arguments{
\item{post.samples}{an object of class \code{draws_df}, \code{draws_matrix}, \code{matrix}, or \code{data.frame} giving posterior
samples of a GLM under NAPP, such as the output from \code{\link[=glm.napp]{glm.napp()}}. Each row corresponds to
the posterior samples obtained from one iteration of MCMC. The column names of \code{post.samples}
should include the names of covariates for regression coefficients, such as "(Intercept)", and
"dispersion" for the dispersion parameter, if applicable.}

\item{formula}{a two-sided formula giving the relationship between the response variable and covariates.}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}.}

\item{data.list}{a list of \code{data.frame}s. The first element in the list is the current data, and the rest
are the historical datasets.}

\item{offset.list}{a list of vectors giving the offsets for each data. The length of offset.list is equal to
the length of data.list. The length of each element of offset.list is equal to the number
of rows in the corresponding element of data.list. Defaults to a list of vectors of 0s.}

\item{a0.shape1}{first shape parameter for the i.i.d. beta prior on a0 vector. When \code{a0.shape1 == 1} and
\code{a0.shape2 == 1}, a uniform prior is used.}

\item{a0.shape2}{second shape parameter for the i.i.d. beta prior on a0 vector. When \code{a0.shape1 == 1} and
\code{a0.shape2 == 1}, a uniform prior is used.}

\item{bridge.args}{a \code{list} giving arguments (other than samples, log_posterior, data, lb, ub) to pass onto
\code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}.}
}
\value{
The function returns a \code{list} with the following objects

\describe{
\item{model}{"NAPP"}

\item{logml}{the estimated logarithm of the marginal likelihood}

\item{bs}{an object of class \code{bridge} or \code{bridge_list} giving the output from \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}}
}
}
\description{
Uses Markov chain Monte Carlo (MCMC) and bridge sampling to estimate the logarithm of the marginal
likelihood of a GLM under NAPP.
}
\details{
This function mirrors the arguments of \code{\link[=glm.napp]{glm.napp()}} (except for those relevant for MCMC sampling),
and introduces two additional arguments: \code{post.samples} and \code{bridge.args}. \code{post.samples} provides
posterior samples from the GLM under NAPP (e.g., the output from \code{\link[=glm.napp]{glm.napp()}}), while \code{bridge.args}
specifies arguments to pass onto \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}} (other than \code{samples},
\code{log_posterior}, \code{data}, \code{lb}, \code{ub}).

It is crucial to ensure that the values assigned to the shared arguments in this function and
\code{\link[=glm.napp]{glm.napp()}} correspond to those used in generating \code{post.samples}.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  data(actg019)
  data(actg036)
  ## take subset for speed purposes
  actg019 = actg019[1:100, ]
  actg036 = actg036[1:50, ]
  data_list = list(currdata = actg019, histdata = actg036)
  formula = cd4 ~ treatment + age + race
  family = poisson('log')
  d.napp = glm.napp(
    formula = formula, family = family,
    data.list = data_list,
    chains = 1, iter_warmup = 500, iter_sampling = 1000
  )
  glm.logml.napp(
    post.samples = d.napp,
    formula = formula, family = family,
    data.list = data_list,
    bridge.args = list(silent = TRUE)
  )
}
}
\references{
Ibrahim, J. G., Chen, M., Gwon, Y., and Chen, F. (2015). The power prior: Theory and applications. Statistics in Medicine, 34(28), 3724–3749.

Gronau, Q. F., Singmann, H., and Wagenmakers, E.-J. (2020). bridgesampling: An r package for estimating normalizing constants. Journal of Statistical Software, 92(10).
}
