% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_napp.R
\name{glm.napp}
\alias{glm.napp}
\title{Normalized asymptotic power prior (NAPP)}
\usage{
glm.napp(
  formula,
  family,
  data,
  histdata,
  a0.shape1 = 1,
  a0.shape2 = 1,
  offset = NULL,
  offset0 = NULL,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}}

\item{data}{a \code{data.frame} giving the current data}

\item{histdata}{a \code{data.frame} giving the historical data}

\item{a0.shape1}{\code{shape1} parameter for a beta prior on the power prior parameter. When \code{a0.shape1 == 1} and \code{a0.shape2 == 1}, a uniform prior is used}

\item{a0.shape2}{\code{shape1} parameter for a beta prior on the power prior parameter. When \code{a0.shape1 == 1} and \code{a0.shape2 == 1}, a uniform prior is used}

\item{offset}{vector whose dimension is equal to the rows of the current data set giving an offset for the current data. Defaults to a vector of 0s}

\item{offset0}{vector whose dimension is equal to the rows of the historical data set giving an offset for the historical data. Defaults to a vector of 0s}

\item{...}{arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (e.g. iter, chains).}
}
\value{
\preformatted{        an object of class `stanfit` giving posterior samples
}
}
\description{
Sample from the posterior distribution of a GLM using the normalized
asymptotic power prior. The regression coefficients and logarithm of the
dispersion parameter are a multivariate normal distribution with mean
equal to the maximum likelihood estimate of the historical data and
covariance matrix equal to \eqn{a0^{-1}} multiplied by the inverse Fisher
information matrix of the historical data, where a0 is the power prior
parameter (treated as random).
}
