% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_npp_lognc.R
\name{glm.npp.lognc}
\alias{glm.npp.lognc}
\title{Estimate the logarithm of the normalizing constant for normalized power prior}
\usage{
glm.npp.lognc(
  formula,
  family,
  histdata,
  a0,
  beta.mean = NULL,
  beta.cov = NULL,
  disp.shape = 2.1,
  disp.scale = 1.1,
  offset0 = NULL,
  bridge.args = NULL,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{stats::family()}}}

\item{histdata}{a \code{data.frame} giving the historical data}

\item{a0}{the power prior parameter (a scalar between 0 and 1)}

\item{beta.mean}{mean parameter for the normal initial prior on the regression coefficients given the dispersion parameter. Defaults to a vector of 0s}

\item{beta.cov}{covariance parameter for the conditional multivariate normal initial prior on the regression coefficients. The covariance used is \code{dispersion * beta.cov}. Defaults to a diagonal matrix of 100s}

\item{disp.shape}{shape parameter for inverse-gamma initial prior on dispersion parameter}

\item{disp.scale}{scale parameter for inverse-gamma initial prior on dispersion parameter}

\item{offset0}{vector whose dimension is equal to the rows of the historical data set giving an offset for the historical data. Defaults to a vector of 0s}

\item{bridge.args}{<optional> a \code{list} giving arguments to pass onto \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}}

\item{...}{arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (e.g. iter, chains)}
}
\value{
\preformatted{        a vector giving the value of a0, the estimated logarithm of the normalizing constant, the minimum effective sample size of
                the MCMC sampling, and the maximum Rhat.
}
}
\description{
Uses Markov chain Monte Carlo and bridge sampling
to estimate the logarithm of the normalizing
constant for the normalized power prior for a fixed value of a0. The discounts
the historical data likelihood by a value a0 between 0 and 1. The initial prior
is a multivariate-normal inverse-gamma prior on the regression coefficients
(conditional on the dispersion) and the dispersion parameter (if applicable).
}
