% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_robustmap.R
\name{glm.robustmap}
\alias{glm.robustmap}
\title{Robust MAP Prior}
\usage{
glm.robustmap(
  formula,
  family,
  data,
  histdata,
  w = 0.1,
  norm.hp.mean = NULL,
  norm.hp.cov = NULL,
  iw.hp.df = NULL,
  iw.hp.scale = NULL,
  norm.vague.mean = NULL,
  norm.vague.cov = NULL,
  disp.shape = 2.1,
  disp.scale = 1.1,
  disp0.shape = 2.1,
  disp0.scale = 1.1,
  offset = NULL,
  offset0 = NULL,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}}

\item{data}{a \code{data.frame} giving the current data}

\item{histdata}{a \code{data.frame} giving the historical data}

\item{w}{a scalar between 0 and 1 giving how much weight to put on the historical data}

\item{norm.hp.mean}{a vector whose dimension is equal to the number of regression coefficients giving the
mean for the normal hyperprior on the regression coefficients. Defaults to a vector of zeros.}

\item{norm.hp.cov}{a covariance matrix whose dimension is equal to the number of regression coefficients
that gives the covariance for the normal hyperprior on the regression coefficients.
Defaults to a diagonal matrix of 1s.}

\item{iw.hp.df}{degrees of freedom for the inverse-wishart prior on the covariance
matrix of the regression coefficients. Defaults to \code{num.predictors + 10}.}

\item{iw.hp.scale}{scale matrix for inverse-Wishart prior on the covariance of the regression coefficients
Defaults to a diagonal matrix of 1s}

\item{norm.vague.mean}{mean hyperparameter for vague normal prior on regression coefficients (second part of mixture). Defaults to a vector of 0s}

\item{norm.vague.cov}{covariance hyperparameter for vague normal prior on regression coefficients (second part of mixture). Defaults to a diagonal matrix of 100s}

\item{disp.shape}{shape parameter for inverse-gamma prior on dispersion parameter for current data set}

\item{disp.scale}{scale parameter for inverse-gamma prior on dispersion parameter for current data set}

\item{disp0.shape}{shape parameter for inverse-gamma prior on dispersion parameter for historical data set}

\item{disp0.scale}{scale parameter for inverse-gamma prior on dispersion parameter for historical data set}

\item{offset}{vector whose dimension is equal to the rows of the current data set giving an offset for the current data. Defaults to a vector of 0s}

\item{offset0}{vector whose dimension is equal to the rows of the historical data set giving an offset for the historical data. Defaults to a vector of 0s}

\item{...}{arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (e.g. iter, chains).}
}
\value{
\preformatted{           an object of class `stanfit` giving posterior samples
}
}
\description{
Sample from the posterior distribution of a GLM using the Robust Meta-Analytic
Predictive (MAP) Prior Of Schmidli. This prior is very similar to the
Bayesian hierarchical model (BHM), except the prior is a mixture prior where
one component is the BHM and the second is vague (noninformative) prior.
}
\examples{
data(actg019)
data(actg036)
## take subset for speed purposes
actg019 = actg019[1:100, ]
actg036 = actg036[1:50, ]
glm.robustmap(
  cd4 ~ treatment + age + race,
  family = poisson(), data = actg019, histdata = actg036,
  w = 0.1,
  chains = 1, warmup = 1000, iter = 2000
)

}
