% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfam_loglik.R
\name{glm_lp}
\alias{glm_lp}
\title{wrapper function compute density for a given link function and a given distribution}
\usage{
glm_lp(y, beta, X, dist, link, offs, phi)
}
\arguments{
\item{y}{response vector}

\item{beta}{regression coefficients}

\item{X}{design matrix}

\item{dist}{integer giving distribution}

\item{link}{integer giving link function}

\item{offs}{offset}

\item{phi}{dispersion parameter}
}
\description{
wrapper function compute density for a given link function and a given distribution
}
