% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbayes-package.R
\docType{package}
\name{hdbayes-package}
\alias{hdbayes-package}
\alias{hdbayes}
\title{The 'hdbayes' package.}
\description{
Bayesian analysis of generalized linear models using historical data
}
\concept{help}
