% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_npp.R
\name{lm.npp}
\alias{lm.npp}
\title{Posterior of normalized power prior}
\usage{
lm.npp(
  formula,
  data,
  histdata,
  beta.mean = NULL,
  beta.cov = NULL,
  sigmasq.shape = 2.1,
  sigmasq.scale = 1.1,
  a0.shape1 = 1,
  a0.shape2 = 1,
  offset = NULL,
  offset0 = NULL,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates}

\item{data}{a \code{data.frame} giving the current data}

\item{histdata}{a \code{data.frame} giving the historical data}

\item{beta.mean}{mean parameter for initial prior on regression coefficients (including intercept). Defaults to a vector of zeros.}

\item{beta.cov}{covariance parameter for initial prior on regression coefficients (including intercept). Defaults to a diagonal covariance matrix where each variance is equal to 100.}

\item{sigmasq.shape}{shape parameter for inverse-gamma prior on variance parameter}

\item{sigmasq.scale}{scale parameter for inverse-gamma prior on variance parameter}

\item{a0.shape1}{first shape parameter for beta prior on a0. When \code{a0.shape1 == 1} and \code{a0.shape2 == 1}, a uniform prior is used.}

\item{a0.shape2}{first shape parameter for beta prior on a0. When \code{a0.shape1 == 1} and \code{a0.shape2 == 1}, a uniform prior is used.}

\item{offset}{vector whose dimension is equal to the rows of the current data set giving an offset for the current data. Defaults to a vector of 0s}

\item{offset0}{vector whose dimension is equal to the rows of the historical data set giving an offset for the historical data. Defaults to a vector of 0s}

\item{...}{arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (e.g. iter, chains).}
}
\value{
\preformatted{       an object of class `stanfit` giving posterior samples
}
}
\description{
Sample from the posterior distribution of a normal linear model
using the normalized power prior (NPP)
}
\examples{
data(actg019)
data(actg036)
lm.npp(
   cd4 ~ treatment + age + race,
   data = actg019, histdata = actg036, chains = 1
)


}
