% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfam_loglik.R
\name{normal_glm_lp}
\alias{normal_glm_lp}
\title{compute density for normal GLM}
\usage{
normal_glm_lp(y, beta, X, link, offs, phi)
}
\arguments{
\item{y}{response vector}

\item{beta}{regression coefficients}

\item{X}{design matrix}

\item{link}{integer giving link function}

\item{offs}{offset}

\item{phi}{dispersion parameter (variance)}
}
\description{
compute density for normal GLM
}
