% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfam_loglik.R
\name{poisson_glm_lp}
\alias{poisson_glm_lp}
\title{compute density for poisson GLM}
\usage{
poisson_glm_lp(y, beta, X, link, offs, phi = 1)
}
\arguments{
\item{y}{response vector}

\item{beta}{regression coefficients}

\item{X}{design matrix}

\item{link}{integer giving link function}

\item{offs}{offset}

\item{phi}{dispersion parameter (phi = 1)}
}
\description{
compute density for poisson GLM
}
