% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{to.vector}
\alias{to.vector}
\title{transfer a scalar/vector/NULL into a vector of given length and (default) values}
\usage{
to.vector(param, default.value = 0, len = 1)
}
\arguments{
\item{param}{a scalar or a vector if param is not NULL}

\item{default.value}{default value for param. Defaults to 0.}

\item{len}{length (number of elements) of param. Defaults to 1.}
}
\description{
transfer a scalar/vector/NULL into a vector of given length and (default) values
}
