% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E1694-data.R
\docType{data}
\name{E1694}
\alias{E1694}
\title{ECOG E1694 Data
from patients receiving the vaccine (GMK) only}
\format{
A data frame with 200 rows and 6 variables:
\describe{
\item{failtime}{relapse-free surviva (RFS) times (in months)}
\item{failind}{event indicator, 0 = censored, 1 = failed}
\item{trt}{treatment indicator, 0 = GMK, 1 = IFN}
\item{sex}{gender indicator, 0 = male, 1 = female}
\item{age}{age, in years}
\item{perform}{ECOG performance status indicator, 0 = fully active patient, able to
carry on all pre-disease performance without restriction,
1 = restricted in physically strenuous activity, but are ambulatory
and able to carry out work of a light or sedentary nature}
}
}
\description{
ECOG E1694 Data
from patients receiving the vaccine (GMK) only
}
\references{
DOI: 10.1200/JCO.2001.19.9.2370
}
\author{
Joseph Ibrahim \email{ibrahim@bios.unc.edu}
}
\keyword{data}
