% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_bhm.R
\name{glm.bhm}
\alias{glm.bhm}
\title{Bayesian hierarchical model}
\usage{
glm.bhm(
  formula,
  family,
  data,
  histdata,
  norm.hp.mean = NULL,
  norm.hp.cov = NULL,
  iw.hp.df = NULL,
  iw.hp.scale = NULL,
  disp.shape = 2.1,
  disp.scale = 1.1,
  disp0.shape = 2.1,
  disp0.scale = 1.1,
  offset = NULL,
  offset0 = NULL,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}}

\item{data}{a \code{data.frame} giving the current data}

\item{histdata}{a \code{data.frame} giving the historical data}

\item{norm.hp.mean}{a vector whose dimension is equal to the number of regression coefficients giving the
mean for the normal hyperprior on the regression coefficients. Defaults to a vector of zeros.}

\item{norm.hp.cov}{a covariance matrix whose dimension is equal to the number of regression coefficients
that gives the covariance for the normal hyperprior on the regression coefficients.
Defaults to a diagonal matrix of 1s.}

\item{iw.hp.df}{degrees of freedom for the inverse-wishart prior on the covariance
matrix of the regression coefficients. Defaults to \code{num.predictors + 10}.}

\item{iw.hp.scale}{scale matrix for inverse-Wishart prior on the covariance of the regression coefficients
Defaults to a diagonal matrix of 1s.}

\item{disp.shape}{shape parameter for inverse-gamma prior on dispersion parameter for current data set}

\item{disp.scale}{scale parameter for inverse-gamma prior on dispersion parameter for current data set}

\item{disp0.shape}{shape parameter for inverse-gamma prior on dispersion parameter for historical data set}

\item{disp0.scale}{scale parameter for inverse-gamma prior on dispersion parameter for historical data set}

\item{offset}{vector whose dimension is equal to the rows of the current data set giving an offset for the current data. Defaults to a vector of 0s}

\item{offset0}{vector whose dimension is equal to the rows of the historical data set giving an offset for the historical data. Defaults to a vector of 0s}

\item{...}{arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (e.g. iter, chains).}
}
\value{
\preformatted{        an object of class `stanfit` giving posterior samples
}
}
\description{
Sample from the posterior distribution of a GLM using the Bayesian Hierarchical
Model (BHM). This model assumes that the regression coefficients of the
historical and current data are different, but are correlated through a
common distribution, whose hyperparameters (i.e., mean and covariance) are treated as random.
The mean hyperparmater is multivariate normal and the covariance hyperparameter is
inverse-Wishart.
The dispersion parameters (if applicable) are independent.
}
\examples{
data(actg019)
data(actg036)
## take subset for speed purposes
actg019 = actg019[1:200, ]
actg036 = actg036[1:100, ]
glm.bhm(
  cd4 ~ treatment + age + race,
  family = poisson(), data = actg019, histdata = actg036,
  chains = 1, warmup = 500, iter = 1000
)
}
