% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_commensurate.R
\name{glm.commensurate}
\alias{glm.commensurate}
\title{Commensurate prior}
\usage{
glm.commensurate(
  formula,
  family,
  data,
  histdata,
  tau,
  beta0.mean = NULL,
  beta0.cov = NULL,
  disp.shape = 2.1,
  disp.scale = 1.1,
  disp0.shape = 2.1,
  disp0.scale = 1.1,
  offset = NULL,
  offset0 = NULL,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}}

\item{data}{a \code{data.frame} giving the current data}

\item{histdata}{a \code{data.frame} giving the historical data}

\item{tau}{a vector giving the commensurate prior parameter. The
dimension must equal the number of predictors (including the intercept). Each
element must be positive, corresponding to a normal precision
parameter.}

\item{beta0.mean}{mean parameter for initial prior on regression coefficients (including intercept). Defaults to a vector of zeros.}

\item{beta0.cov}{covariance parameter for initial prior on regression coefficients (including intercept). Defaults to a diagonal covariance matrix where each variance is equal to 100.}

\item{disp.shape}{shape parameter for inverse-gamma prior on dispersion parameter for current data set}

\item{disp.scale}{scale parameter for inverse-gamma prior on dispersion parameter for current data set}

\item{disp0.shape}{shape parameter for inverse-gamma prior on dispersion parameter for historical data set}

\item{disp0.scale}{scale parameter for inverse-gamma prior on dispersion parameter for historical data set}

\item{offset}{vector whose dimension is equal to the rows of the current data set giving an offset for the current data. Defaults to a vector of 0s}

\item{offset0}{vector whose dimension is equal to the rows of the historical data set giving an offset for the historical data. Defaults to a vector of 0s}

\item{...}{arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (e.g. iter, chains).}
}
\value{
\preformatted{       an object of class `stanfit` giving posterior samples
}
}
\description{
Sample from the posterior distribution of a GLM using the Commensurate
prior of Hobbs et al. This prior assumes that the regression coefficients
for the current data set conditional on those for the historical data set are
multivariate normal with mean equal to the regression coefficients of the historical data
and covariance equal to the inverse of a user-specified precision parameter (tau).
Dispersion parameters (if applicable) between the
current and historical data sets are treated as independent.
}
\examples{
data(actg019)
data(actg036)
## take subset for speed purposes
actg019 = actg019[1:100, ]
actg036 = actg036[1:50, ]
glm.commensurate(
  cd4 ~ treatment + age + race,
  tau = rep(5, 4),    ## 4 parameters including intercept
  family = poisson(), data = actg019, histdata = actg036,
  chains = 1, warmup = 500, iter = 1000
)
}
