% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_leap_lognc.R
\name{glm.leap.lognc}
\alias{glm.leap.lognc}
\title{Estimate the logarithm of the normalizing constant for latent exchangeability prior (LEAP)}
\usage{
glm.leap.lognc(post.samples, bridge.args = NULL)
}
\arguments{
\item{post.samples}{output from \code{\link[=glm.leap]{glm.leap()}} giving posterior samples of a GLM under the latent exchangeability
prior (LEAP), with an attribute called 'data' which includes the list of variables specified
in the data block of the Stan program.}

\item{bridge.args}{a \code{list} giving arguments (other than \code{samples}, \code{log_posterior}, \code{data}, \code{lb}, and \code{ub})
to pass onto \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}.}
}
\value{
The function returns a \code{list} with the following objects

\describe{
\item{lognc}{the estimated logarithm of the normalizing constant}

\item{bs}{an object of class \code{bridge} or \code{bridge_list} giving the output from \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}}
}
}
\description{
Uses bridge sampling to estimate the logarithm of the normalizing constant for the latent exchangeability
prior (LEAP) using historical data sets.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  data(actg036)
  ## take subset for speed purposes
  actg036   = actg036[1:50, ]
  formula   = outcome ~ scale(age) + race + treatment + scale(cd4)
  family    = binomial('logit')
  hist.data.list = list(histdata = actg036)
  hist.standat = hdbayes:::get.stan.data.leap(
    formula = formula,
    family = family,
    data.list = hist.data.list,
    K = 2,
    all.hist = T
  )
  glm_leap_prior = instantiate::stan_package_model(
    name = "glm_leap_prior",
    package = "hdbayes"
  )
  fit = glm_leap_prior$sample(
    data = hist.standat,
    iter_warmup = 500, iter_sampling = 1000, chains = 1
  )
  d.leap.hist  = fit$draws(format = 'draws_df')
  attr(x = d.leap.hist, which = 'data') = hist.standat
  glm.leap.lognc(
    post.samples = d.leap.hist
  )
}
}
\references{
Alt, E. M., Chang, X., Jiang, X., Liu, Q., Mo, M., Xia, H. M., and Ibrahim, J. G. (2023). LEAP: The latent exchangeability prior for borrowing information from historical data. arXiv preprint.

Gronau, Q. F., Singmann, H., and Wagenmakers, E.-J. (2020). bridgesampling: An r package for estimating normalizing constants. Journal of Statistical Software, 92(10).
}
