% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_leap.R
\name{glm.leap}
\alias{glm.leap}
\title{Posterior of LEAP}
\usage{
glm.leap(
  formula,
  family,
  data.list,
  K = 2,
  beta.mean = NULL,
  beta.cov = NULL,
  disp.shape = 2.1,
  disp.scale = 1.1,
  offset.list = NULL,
  local.location = NULL,
  iter_warmup = 1000,
  iter_sampling = 1000,
  chains = 4,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}}

\item{data.list}{a list of \code{data.frame} giving the current data followed by historical data}

\item{K}{the desired number of classes to identify}

\item{beta.mean}{a \verb{p x K} matrix of mean parameter for initial prior on regression coefficients (including intercept). Defaults to a matrix of zeros.}

\item{beta.cov}{a list of \code{K} \code{data.frame} each size \verb{p x p} giving covariance parameter for initial prior on regression coefficients (including intercept). Each defaults to a diagonal covariance matrix where each variance is equal to 100.}

\item{disp.shape}{shape parameter for inverse-gamma prior on dispersion parameter}

\item{disp.scale}{scale parameter for inverse-gamma prior on dispersion parameter}

\item{offset.list}{a list of \code{data.frame} giving the offset for current data followed by historical data. For each \code{data.frame}, the number of rows correpond to observations and columns correspond to classes. Defaults to matrices of 0s}

\item{local.location}{a file path giving the desired location of the local copies of all the .stan model files in the
package. Defaults to the path created by \code{rappdirs::user_cache_dir("hdbayes")}.}

\item{iter_warmup}{number of warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_warmup} in
\code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{iter_sampling}{number of post-warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_sampling}
in \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{chains}{number of Markov chains to run. Defaults to 4. See the argument \code{chains} in \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{...}{arguments passed to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}} (e.g. seed, refresh, init).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{             an object of class `draws_df` giving posterior samples
}\if{html}{\out{</div>}}
}
\description{
...
}
\examples{
data(actg019)
data(actg036)
# take subset for speed purposes
actg019 = actg019[1:100, ]
actg036 = actg036[1:50, ]
glm.leap(
  cd4 ~ treatment + age + race,
  family = poisson(),
  data.list = list(actg019, actg036),
  K = 2,
  chains = 1, iter_warmup = 500, iter_sampling = 1000
)

}
