% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_logml_reference.R
\name{glm.logml.reference}
\alias{glm.logml.reference}
\title{Log marginal likelihood of a GLM under a non-informative reference prior}
\usage{
glm.logml.reference(
  post.samples,
  formula,
  family,
  data.list,
  offset.list = NULL,
  beta.mean = NULL,
  beta.sd = NULL,
  disp.mean = NULL,
  disp.sd = NULL,
  bridge.args = NULL
)
}
\arguments{
\item{post.samples}{an object of class \code{draws_df}, \code{draws_matrix}, \code{matrix}, or \code{data.frame} giving posterior
samples of a GLM under a Ref prior, such as the output from \code{\link[=glm.reference]{glm.reference()}}. Each row
corresponds to the posterior samples obtained from one iteration of MCMC. The column names
of \code{post.samples} should include the names of covariates for regression coefficients, such
as "(Intercept)", and "dispersion" for the dispersion parameter, if applicable.}

\item{formula}{a two-sided formula giving the relationship between the response variable and covariates.}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}.}

\item{data.list}{a list consisting of only one \code{data.frame} containing the current data.}

\item{offset.list}{a list consisting of only one vector giving the offset for the current data. The length of
the vector is equal to the number of rows in the current data. The vector has all values
set to 0 by default.}

\item{beta.mean}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the mean parameters for the normal prior on regression coefficients. If a scalar is provided,
beta.mean will be a vector of repeated elements of the given scalar. Defaults to a vector of 0s.}

\item{beta.sd}{a scalar or a vector whose dimension is equal to the number of regression coefficients giving
the sd parameters for the normal prior on regression coefficients. If a scalar is provided,
same as for beta.mean. Defaults to a vector of 10s.}

\item{disp.mean}{location parameter for the half-normal prior on dispersion parameter. Defaults to 0.}

\item{disp.sd}{scale parameter for the half-normal prior on dispersion parameter. Defaults to 10.}

\item{bridge.args}{a \code{list} giving arguments (other than samples, log_posterior, data, lb, ub) to pass
onto \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}.}
}
\value{
The function returns a \code{list} with the following objects

\describe{
\item{model}{"Reference"}

\item{logml}{the estimated logarithm of the marginal likelihood}

\item{bs}{an object of class \code{bridge} or \code{bridge_list} giving the output from \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}}
}
}
\description{
Uses Markov chain Monte Carlo (MCMC) and bridge sampling to estimate the logarithm of the marginal
likelihood of a GLM under a non-informative reference prior (referred to as Ref prior).
}
\details{
This function mirrors the same arguments as \code{\link[=glm.reference]{glm.reference()}} (except for those relevant for MCMC sampling),
and introduces two additional parameters: \code{post.samples} and \code{bridge.args}. \code{post.samples} provides posterior
samples generated from the GLM under a Ref prior (e.g., the output from \code{\link[=glm.reference]{glm.reference()}}), whereas
\code{bridge.args} specifies arguments to pass onto \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}} (other than \code{samples},
\code{log_posterior}, \code{data}, \code{lb}, and \code{ub}).

It is important to ensure that the values assigned to the shared arguments in this function and
\code{\link[=glm.reference]{glm.reference()}} align with those used in generating \code{post.samples}.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  data(actg019)
  actg019 = actg019[1:100, ]
  data.list = list(currdata = actg019)
  formula = cd4 ~ treatment + age + race
  family = poisson('log')
  d.ref = glm.reference(
    formula = formula, family = family,
    data.list = data.list,
    chains = 1, iter_warmup = 500, iter_sampling = 1000
  )
  glm.logml.reference(
    post.samples = d.ref,
    formula = formula, family = family,
    data.list = data.list,
    bridge.args = list(silent = TRUE)
  )
}
}
\references{
Gronau, Q. F., Singmann, H., and Wagenmakers, E.-J. (2020). bridgesampling: An r package for estimating normalizing constants. Journal of Statistical Software, 92(10).
}
