% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_napp.R
\name{glm.napp}
\alias{glm.napp}
\title{Normalized asymptotic power prior (NAPP)}
\usage{
glm.napp(
  formula,
  family,
  data.list,
  offset.list = NULL,
  a0.shape1 = 1,
  a0.shape2 = 1,
  local.location = NULL,
  iter_warmup = 1000,
  iter_sampling = 1000,
  chains = 4,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates.}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}.}

\item{data.list}{a list of \code{data.frame}s. The first element in the list is the current data, and the rest
are the historical datasets.}

\item{offset.list}{a list of vectors giving the offsets for each data. The length of offset.list is equal to
the length of data.list. The length of each element of offset.list is equal to the number
of rows in the corresponding element of data.list. Defaults to a list of vectors of 0s.}

\item{a0.shape1}{first shape parameter for the i.i.d. beta prior on a0 vector. When \code{a0.shape1 == 1} and
\code{a0.shape2 == 1}, a uniform prior is used.}

\item{a0.shape2}{second shape parameter for the i.i.d. beta prior on a0 vector. When \code{a0.shape1 == 1} and
\code{a0.shape2 == 1}, a uniform prior is used.}

\item{local.location}{a file path giving the desired location of the local copies of all the .stan model files in the
package. Defaults to the path created by \code{rappdirs::user_cache_dir("hdbayes")}.}

\item{iter_warmup}{number of warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_warmup} in
\code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{iter_sampling}{number of post-warmup iterations to run per chain. Defaults to 1000. See the argument \code{iter_sampling}
in \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{chains}{number of Markov chains to run. Defaults to 4. See the argument \code{chains} in \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{...}{arguments passed to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}} (e.g. seed, refresh, init).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{             an object of class `draws_df` giving posterior samples
}\if{html}{\out{</div>}}
}
\description{
Sample from the posterior distribution of a GLM using the normalized
asymptotic power prior. The regression coefficients and logarithm of the
dispersion parameter are a multivariate normal distribution with mean
equal to the maximum likelihood estimate of the historical data and
covariance matrix equal to \eqn{a0^{-1}} multiplied by the inverse Fisher
information matrix of the historical data, where a0 is the power prior
parameter (treated as random).
}
\examples{
data(actg019)
data(actg036)
## take subset for speed purposes
actg019 = actg019[1:100, ]
actg036 = actg036[1:50, ]
data.list = list(data = actg019, histdata = actg036)
glm.napp(
  cd4 ~ treatment + age + race,
  family = poisson(), data.list = data.list,
  chains = 1, iter_warmup = 500, iter_sampling = 1000
)
}
