% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_pp.R
\name{glm.pp}
\alias{glm.pp}
\title{Posterior of power prior with fixed a0}
\usage{
glm.pp(
  formula,
  family,
  data,
  histdata,
  a0,
  beta.mean = NULL,
  beta.cov = NULL,
  disp.shape = 2.1,
  disp.scale = 1.1,
  offset = NULL,
  offset0 = NULL,
  ...
)
}
\arguments{
\item{formula}{a two-sided formula giving the relationship between the response variable and covariates}

\item{family}{an object of class \code{family}. See \code{\link[stats:family]{?stats::family}}}

\item{data}{a \code{data.frame} giving the current data}

\item{histdata}{a \code{data.frame} giving the historical data}

\item{a0}{scalar between 0 and 1 giving the (fixed) power prior parameter}

\item{beta.mean}{mean parameter for initial prior on regression coefficients (including intercept). Defaults to a vector of zeros.}

\item{beta.cov}{covariance parameter for initial prior on regression coefficients (including intercept). Defaults to a diagonal covariance matrix where each variance is equal to 100.}

\item{disp.shape}{shape parameter for inverse-gamma prior on dispersion parameter}

\item{disp.scale}{scale parameter for inverse-gamma prior on dispersion parameter}

\item{offset}{vector whose dimension is equal to the rows of the current data set giving an offset for the current data. Defaults to a vector of 0s}

\item{offset0}{vector whose dimension is equal to the rows of the historical data set giving an offset for the historical data. Defaults to a vector of 0s}

\item{...}{arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (e.g. iter, chains).}
}
\value{
\preformatted{       an object of class `stanfit` giving posterior samples
}
}
\description{
Sample from the posterior distribution of a GLM using the power prior with
a fixed power prior parameter (a0)
}
\examples{
data(actg019)
data(actg036)
## take subset for speed purposes
actg019 = actg019[1:100, ]
actg036 = actg036[1:50, ]
glm.pp(
  cd4 ~ treatment + age + race,
  family = poisson(), data = actg019, histdata = actg036,
  a0 = 0.5,
  chains = 1, warmup = 500, iter = 1000
)



}
